[Spec 0602][Phase 1b] Extract codev-core package from tower-client
Create @cluesmith/codev-core with subpath exports:
- tower-client: TowerClient class with injectable auth
- auth: readLocalKey (read-only) + ensureLocalKey (CLI-only)
- workspace: encodeWorkspacePath / decodeWorkspacePath
- constants: DEFAULT_TOWER_PORT, AGENT_FARM_DIR
- escape-buffer: EscapeBuffer (browser-safe)

Original tower-client.ts replaced with thin re-exports.
Constructor accepts both number (backward compat) and options
object (injectable getAuthKey for VS Code extension).
