+++ .gitignore
@@ -28,6 +28,7 @@ node_modules/
 packages/codev/dist/
 packages/codev/skeleton/
 packages/types/dist/
+packages/core/dist/
 packages/vscode/dist/
 packages/vscode/out/
 *.tsbuildinfo
